requirements.txt
absl-py
addict
aiohappyeyeballs
aiohttp
aiosignal
albumentations
aliyun-python-sdk-core
aliyun-python-sdk-kms
altair
antlr4-python3-runtime
anyio
astor
astunparse
async-timeout
attrs
basicsr
beautifulsoup4
blinker
braceexpand
cffi
cmake
crcmod
cryptography
datasets
diffusers
dill
easydict
einops
entrypoints
fastapi
fire
flatbuffers
frozenlist
fsspec
ftfy
future
gast
gitdb
gitpython
google-auth-oauthlib
google-pasta
gradio
gradio-client
h5py
imageio
imageio-ffmpeg
imgaug
importlib-metadata
jieba
jmespath
jsonschema
jsonschema-specifications
keras
kornia
langid
lazy-loader
libclang
lit
lxml
markdown
mock
modelscope
multidict
multiprocess
mypy-extensions
numpy
oauthlib
omegaconf
open-clip-torch
opencv-contrib-python
opencv-python
opt-einsum
oss2
pandas
pillow
protobuf
pycparser
pycryptodome
pydantic
pydeck
pydeprecate
pympler
pyre-extensions
python-docx
pytorch-lightning
pywavelets
rapidfuzz
referencing
requests
requests-oauthlib
rpds-py
sacremoses
safetensors
scikit-image
semver
simplejson
smmap
sortedcontainers
soupsieve
starlette
subword-nmt
tensorboard
tensorboard-data-server
tensorflow
tensorflow-estimator
tensorflow-io-gcs-filesystem
termcolor
test-tube
tifffile
timm
tokenizers
toml
torch
torchmetrics
torchvision
tqdm
transformers
triton
typing-extensions
typing-inspect
tzlocal
ujson
validators
watchdog
webdataset
werkzeug
wrapt
xformers
xxhash
yapf
yarl
zipp